api-gateway/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
